common.sh: Add newer smfs dump sysfs path

+++ common.sh
@@ -1870,7 +1870,11 @@ function dmfs_dump() {
 
 function smfs_dump() {
     local dump=${1:-dump}
-    cat /proc/driver/mlx5_core/smfs_dump/fdb/$PCI > /tmp/$dump || err "smfs dump failed"
+    if [ -f /sys/kernel/debug/mlx5/$PCI/steering/fdb ]; then
+        cat /sys/kernel/debug/mlx5/$PCI/steering/fdb/* > /tmp/$dump || err "smfs dump failed"
+    else
+        cat /proc/driver/mlx5_core/smfs_dump/fdb/$PCI > /tmp/$dump || err "smfs dump failed"
+    fi
 }
 
 function fw_dump() {
